refactor: simplify entity services
