+++ src/JhipsterSampleApplication/Controllers/BirthdaysController.cs
@@ -6,15 +6,13 @@ using Nest;
 using JhipsterSampleApplication.Domain.Services.Interfaces;
 using JhipsterSampleApplication.Domain.Entities;
 using System.Collections.Generic;
-using JHipsterNet.Core.Pagination;
-using Microsoft.AspNetCore.Authorization;
 using Newtonsoft.Json.Linq;
 using System.Linq;
 using System.IO;
 using JhipsterSampleApplication.Dto;
 using Microsoft.Extensions.Logging;
 using AutoMapper;
-using System.Collections.ObjectModel;
+using System.Net;
 
 namespace JhipsterSampleApplication.Controllers
 {
@@ -28,6 +26,7 @@ namespace JhipsterSampleApplication.Controllers
         private readonly ILogger<BirthdaysController> _log;
         private readonly IMapper _mapper;
         private readonly IHistoryService _historyService;
+        private readonly IViewService _viewService;
 
         public BirthdaysController(
             IBirthdayService birthdayService,
@@ -35,7 +34,8 @@ namespace JhipsterSampleApplication.Controllers
             IBqlService<Birthday> bqlService,
             ILogger<BirthdaysController> log,
             IMapper mapper,
-            IHistoryService historyService)
+            IHistoryService historyService,
+            IViewService viewService)
         {
             _birthdayService = birthdayService;
             _elasticClient = elasticClient;
@@ -43,6 +43,7 @@ namespace JhipsterSampleApplication.Controllers
             _log = log;
             _mapper = mapper;
             _historyService = historyService;
+            _viewService = viewService;
         }
 
         public class RawSearchRequestDto
@@ -65,11 +66,31 @@ namespace JhipsterSampleApplication.Controllers
         [Produces("text/html")]
         public async Task<IActionResult> GetHtmlById(string id)
         {
-            var html = await _birthdayService.GetHtmlByIdAsync(id);
-            if (html == null)
+            var searchRequest = new SearchRequest<Birthday>
+            {
+                Query = new QueryContainerDescriptor<Birthday>().Term(t => t.Field("_id").Value(id))
+            };
+            var response = await _birthdayService.SearchAsync(searchRequest, includeDetails: true, pitId: "");
+            if (!response.IsValid || !response.Documents.Any())
             {
                 return NotFound();
             }
+            var birthday = response.Documents.First();
+            var fullName = ($"{birthday.Fname} {birthday.Lname}").Trim();
+            var wikipediaHtml = birthday.Wikipedia ?? string.Empty;
+            var title = string.IsNullOrWhiteSpace(fullName) ? "Birthday" : fullName;
+            var html = "<!doctype html>" +
+                       "<html><head>" +
+                       "<meta charset=\"utf-8\">" +
+                       "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1\">" +
+                       "<base target=\"_blank\">" +
+                       "<title>" + WebUtility.HtmlEncode(title) + "</title>" +
+                       "<style>body{margin:0;padding:8px;font-family:system-ui,-apple-system,Segoe UI,Roboto,Ubuntu,Cantarell,Noto Sans,Helvetica Neue,Arial,\"Apple Color Emoji\",\"Segoe UI Emoji\";font-size:14px;line-height:1.4;color:#111} .empty{color:#666}</style>" +
+                       "</head><body>" +
+                       (string.IsNullOrWhiteSpace(wikipediaHtml)
+                           ? ("<div class=\"empty\">No Wikipedia content available." + (string.IsNullOrWhiteSpace(fullName) ? string.Empty : (" for " + WebUtility.HtmlEncode(fullName))) + "</div>")
+                           : wikipediaHtml)
+                       + "</body></html>";
             return Content(html, "text/html");
         }
 
@@ -117,7 +138,7 @@ namespace JhipsterSampleApplication.Controllers
             JObject queryObject = new JObject(
                 new JProperty("query_string", queryStringObject)
             );
-            var result = await _birthdayService.Search(queryObject, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
+            var result = await PerformSearch(queryObject, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
             return Ok(result);
         }
 
@@ -141,7 +162,7 @@ namespace JhipsterSampleApplication.Controllers
         {
             var ruleset = _mapper.Map<Ruleset>(rulesetDto);
             var queryObject = await _birthdayService.ConvertRulesetToElasticSearch(ruleset);
-            var result = await _birthdayService.Search(queryObject, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
+            var result = await PerformSearch(queryObject, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
             return Ok(result);
         }
 
@@ -163,7 +184,7 @@ namespace JhipsterSampleApplication.Controllers
             [FromQuery] string? pitId = null,
             [FromQuery] string[]? searchAfter = null)
         {
-            var result = await _birthdayService.Search(elasticsearchQuery, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
+            var result = await PerformSearch(elasticsearchQuery, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
             return Ok(result);
         }
         /// <summary>
@@ -193,20 +214,263 @@ namespace JhipsterSampleApplication.Controllers
             var ruleset = _mapper.Map<Ruleset>(rulesetDto);
             var queryObject = await _birthdayService.ConvertRulesetToElasticSearch(ruleset);
             await _historyService.Save(new History { User = User?.Identity?.Name, Domain = "birthday", Text = bqlQuery });
-            var result = await _birthdayService.Search(queryObject, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
+            var result = await PerformSearch(queryObject, pageSize, from, sort, includeDetails, view, category, secondaryCategory, pitId, searchAfter);
             return Ok(result);
+        }
+
+        private async Task<object> PerformSearch(JObject elasticsearchQuery, int pageSize = 20, int from = 0, string? sort = null,
+            bool includeDetails = false, string? view = null, string? category = null, string? secondaryCategory = null,
+            string? pitId = null, string[]? searchAfter = null)
+        {
+            bool isHitFromViewDrilldown = false;
+            if (!string.IsNullOrEmpty(view))
+            {
+                var viewDto = await _viewService.GetByIdAsync(view) ?? throw new ArgumentException($"view '{view}' not found");
+                if (category == null)
+                {
+                    if (secondaryCategory != null)
+                    {
+                        throw new ArgumentException($"secondaryCategory '{secondaryCategory}' should be null because category is null");
+                    }
+                    try
+                    {
+                        var viewResult = await _birthdayService.SearchWithElasticQueryAndViewAsync(elasticsearchQuery, viewDto, pageSize, from);
+                        return new SearchResultDto<ViewResultDto> { Hits = viewResult, HitType = "view", ViewName = view };
+                    }
+                    catch
+                    {
+                        var baseField = (viewDto.Aggregation ?? string.Empty).Replace(".keyword", string.Empty);
+                        var missingQuery = new JObject(
+                            new JProperty("bool", new JObject(
+                                new JProperty("must", new JArray(
+                                    elasticsearchQuery,
+                                    new JObject(
+                                        new JProperty("bool", new JObject(
+                                            new JProperty("must_not", new JArray(
+                                                new JObject(new JProperty("exists", new JObject(new JProperty("field", baseField))))
+                                            ))
+                                        ))
+                                    )
+                                ))
+                            ))
+                        );
+                        var missingSearch = new SearchRequest<Birthday>
+                        {
+                            Size = 0,
+                            Query = new QueryContainerDescriptor<Birthday>().Raw(missingQuery.ToString())
+                        };
+                        var missResp = await _birthdayService.SearchAsync(missingSearch, includeDetails: false, pitId: "");
+                        long missingCount = missResp.IsValid ? missResp.Total : 0;
+                        var hits = new List<ViewResultDto>();
+                        if (missingCount > 0)
+                        {
+                            hits.Add(new ViewResultDto { CategoryName = "(Uncategorized)", Count = missingCount, NotCategorized = true });
+                        }
+                        return new SearchResultDto<ViewResultDto> { Hits = hits, HitType = "view", ViewName = view };
+                    }
+                }
+                if (category == "(Uncategorized)")
+                {
+                    var baseField = (viewDto.Aggregation ?? string.Empty).Replace(".keyword", string.Empty);
+                    var missingFilter = new JObject(
+                        new JProperty("bool", new JObject(
+                            new JProperty("should", new JArray(
+                                new JObject(
+                                    new JProperty("bool", new JObject(
+                                        new JProperty("must_not", new JArray(
+                                            new JObject(
+                                                new JProperty("exists", new JObject(
+                                                    new JProperty("field", baseField)
+                                                ))
+                                            )
+                                        ))
+                                    ))
+                                ),
+                                new JObject(
+                                    new JProperty("term", new JObject(
+                                        new JProperty(viewDto.Aggregation ?? string.Empty, "")
+                                    ))
+                                )
+                            )),
+                            new JProperty("minimum_should_match", 1)
+                        ))
+                    );
+                    elasticsearchQuery = new JObject(
+                        new JProperty("bool", new JObject(
+                            new JProperty("must", new JArray(
+                                missingFilter,
+                                elasticsearchQuery
+                            ))
+                        ))
+                    );
+                }
+                else
+                {
+                    string categoryQuery = string.IsNullOrEmpty(viewDto.CategoryQuery) ? $"{viewDto.Aggregation}:\"{category}\"" : viewDto.CategoryQuery.Replace("{}", category);
+                    elasticsearchQuery = new JObject(
+                        new JProperty("bool", new JObject(
+                            new JProperty("must", new JArray(
+                                new JObject(
+                                    new JProperty("query_string", new JObject(
+                                        new JProperty("query", categoryQuery)
+                                    ))
+                                ),
+                                elasticsearchQuery
+                            ))
+                        ))
+                    );
+                }
+                var secondaryViewDto = await _viewService.GetChildByParentIdAsync(view);
+                if (secondaryViewDto != null)
+                {
+                    if (secondaryCategory == null)
+                    {
+                        var viewSecondaryResult = await _birthdayService.SearchWithElasticQueryAndViewAsync(elasticsearchQuery, secondaryViewDto, pageSize, from);
+                        return new SearchResultDto<ViewResultDto> { Hits = viewSecondaryResult, HitType = "view", ViewName = view, viewCategory = category };
+                    }
+                    if (secondaryCategory == "(Uncategorized)")
+                    {
+                        isHitFromViewDrilldown = true;
+                        var secondaryBaseField = (secondaryViewDto.Aggregation ?? string.Empty).Replace(".keyword", string.Empty);
+                        var secondaryMissing = new JObject(
+                            new JProperty("bool", new JObject(
+                                new JProperty("should", new JArray(
+                                    new JObject(
+                                        new JProperty("bool", new JObject(
+                                            new JProperty("must_not", new JArray(
+                                                new JObject(
+                                                    new JProperty("exists", new JObject(
+                                                        new JProperty("field", secondaryBaseField)
+                                                    ))
+                                                )
+                                            ))
+                                        ))
+                                    ),
+                                    new JObject(
+                                        new JProperty("term", new JObject(
+                                            new JProperty(secondaryViewDto.Aggregation ?? string.Empty, "")
+                                        ))
+                                    )
+                                )),
+                                new JProperty("minimum_should_match", 1)
+                            ))
+                        );
+                        elasticsearchQuery = new JObject(
+                            new JProperty("bool", new JObject(
+                                new JProperty("must", new JArray(
+                                    secondaryMissing,
+                                    elasticsearchQuery
+                                ))
+                            ))
+                        );
+                    }
+                    else
+                    {
+                        isHitFromViewDrilldown = true;
+                        string secondaryCategoryQuery = string.IsNullOrEmpty(secondaryViewDto.CategoryQuery) ? $"{secondaryViewDto.Aggregation}:\"{secondaryCategory}\"" : secondaryViewDto.CategoryQuery.Replace("{}", secondaryCategory);
+                        elasticsearchQuery = new JObject(
+                            new JProperty("bool", new JObject(
+                                new JProperty("must", new JArray(
+                                    new JObject(
+                                        new JProperty("query_string", new JObject(
+                                            new JProperty("query", secondaryCategoryQuery)
+                                        ))
+                                    ),
+                                    elasticsearchQuery
+                                ))
+                            ))
+                        );
+                    }
+                }
+            }
+            var searchRequest = new SearchRequest<Birthday>
+            {
+                Size = pageSize,
+                From = from
+            };
+
+            if (searchAfter != null && searchAfter.Length > 0)
+            {
+                searchRequest.SearchAfter = searchAfter.Cast<object>().ToList();
+                searchRequest.From = null;
+            }
+
+            var sortDescriptor = new List<ISort>();
+            if (!string.IsNullOrEmpty(sort))
+            {
+                var sortParts = sort.Split(':');
+                if (sortParts.Length == 2)
+                {
+                    var field = sortParts[0];
+                    var order = sortParts[1].ToLower() == "desc" ? SortOrder.Descending : SortOrder.Ascending;
+                    sortDescriptor.Add(new FieldSort { Field = field, Order = order });
+                }
+            }
+            sortDescriptor.Add(new FieldSort { Field = "_id", Order = SortOrder.Ascending });
+
+            searchRequest.Query = new QueryContainerDescriptor<Birthday>().Raw(elasticsearchQuery.ToString());
+            searchRequest.Sort = sortDescriptor;
+
+            var response = await _birthdayService.SearchAsync(searchRequest, includeDetails, pitId);
+            var birthdayDtos = new List<BirthdayDto>();
+            foreach (var hit in response.Hits)
+            {
+                var b = hit.Source;
+                var dto = new BirthdayDto
+                {
+                    Id = b.Id!,
+                    Lname = b.Lname,
+                    Fname = b.Fname,
+                    Sign = b.Sign,
+                    Dob = b.Dob,
+                    IsAlive = b.IsAlive,
+                    Text = b.Text,
+                    Wikipedia = includeDetails ? b.Wikipedia : null,
+                    Categories = b.Categories
+                };
+                if (dto.Categories != null && dto.Categories.Count > 0)
+                {
+                    dto.Categories = dto.Categories
+                        .Where(s => !string.IsNullOrWhiteSpace(s))
+                        .Distinct(StringComparer.OrdinalIgnoreCase)
+                        .Select(s => s.Length == 1 ? s.ToUpperInvariant() : s)
+                        .ToList();
+                }
+                birthdayDtos.Add(dto);
+            }
+            List<object>? searchAfterResponse = response.Hits.Count > 0 ? response.Hits.Last().Sorts.ToList() : null;
+            var hitType = isHitFromViewDrilldown ? "hit" : "birthday";
+            return new SearchResultDto<BirthdayDto>
+            {
+                Hits = birthdayDtos,
+                TotalHits = response.Total,
+                HitType = hitType,
+                PitId = searchRequest.PointInTime?.Id,
+                searchAfter = searchAfterResponse
+            };
         }
 
         [HttpGet("{id}")]
         [ProducesResponseType(typeof(BirthdayDto), 200)]
         public async Task<IActionResult> GetById(string id, [FromQuery] bool includeDetails = false)
         {
-            var birthdayDto = await _birthdayService.GetByIdAsync(id, includeDetails);
-            if (birthdayDto == null)
+            var searchRequest = new SearchRequest<Birthday>
+            {
+                Query = new QueryContainerDescriptor<Birthday>().Term(t => t.Field("_id").Value(id)),
+                Source = includeDetails ? null : new SourceFilter { Excludes = new[] { "wikipedia" } }
+            };
+            var response = await _birthdayService.SearchAsync(searchRequest, includeDetails, pitId: "");
+            if (!response.IsValid || !response.Documents.Any())
             {
                 return NotFound();
             }
-            return Ok(birthdayDto);
+            var b = response.Documents.First();
+            var dto = _mapper.Map<BirthdayDto>(b);
+            if (!includeDetails)
+            {
+                dto.Wikipedia = null;
+            }
+            return Ok(dto);
         }
 
         [HttpPost]
